% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.validate_event_time}
\alias{.validate_event_time}
\title{Internal utilities (not exported)}
\usage{
.validate_event_time(time, last_time)
}
\description{
Internal utilities (not exported)
}
\keyword{internal}
