% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{event}
\alias{event}
\title{Feature target: event type}
\usage{
event(event_type)
}
\arguments{
\item{event_type}{Character scalar.}
}
\value{
A target list used by derive().
}
\description{
Create a target describing events of a given type in entity$events.
}
