% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{FileProvider}
\alias{FileProvider}
\title{FileProvider (internal)}
\description{
Loads a bundle from an .rds file. Expects model_spec$path.
}
\details{
Deprecated in v2.0.0. Use \code{\link[=load_model]{load_model()}} or \code{Engine$new(bundle = ...)} instead.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-FileProvider-initialize}{\code{FileProvider$new()}}
    \item \href{#method-FileProvider-load}{\code{FileProvider$load()}}
    \item \href{#method-FileProvider-sample_param_draws}{\code{FileProvider$sample_param_draws()}}
    \item \href{#method-FileProvider-clone}{\code{FileProvider$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileProvider-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-FileProvider-initialize}{}}}
\subsection{\code{FileProvider$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{FileProvider$new(base_path = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{base_path}}{Optional base directory.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileProvider-load"></a>}}
\if{latex}{\out{\hypertarget{method-FileProvider-load}{}}}
\subsection{\code{FileProvider$load()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{FileProvider$load(model_spec, ...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileProvider-sample_param_draws"></a>}}
\if{latex}{\out{\hypertarget{method-FileProvider-sample_param_draws}{}}}
\subsection{\code{FileProvider$sample_param_draws()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{FileProvider$sample_param_draws(model_spec, n_param_draws = 1L, ...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileProvider-clone}{}}}
\subsection{\code{FileProvider$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{FileProvider$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
