\name{merge_patches}
\alias{merge_patches}
\title{Merge two sparse change patches}
\usage{
merge_patches(
  baseline_changes,
  policy_changes,
  merge = c("policy_wins", "baseline_wins", "error_on_conflict")
)
}
\arguments{
\item{baseline_changes}{Named list or NULL.}

\item{policy_changes}{Named list or NULL.}

\item{merge}{Merge strategy.}
}
\value{
Named list or NULL.
}
\description{
Merge two sparse change patches
}
\keyword{internal}
