% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{Engine}
\alias{Engine}
\title{Engine}
\description{
Orchestrates simulation by repeatedly proposing the next event(s), applying a transition
patch, recording the event on a Entity, and stopping when bundle$stop() returns TRUE
(or a max_time / max_events limit is reached).
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Engine-initialize}{\code{Engine$new()}}
    \item \href{#method-Engine-run}{\code{Engine$run()}}
    \item \href{#method-Engine-run_draw}{\code{Engine$run_draw()}}
    \item \href{#method-Engine-clone}{\code{Engine$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Engine-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Engine-initialize}{}}}
\subsection{\code{Engine$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Engine$new(bundle = NULL, ...)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Engine-run"></a>}}
\if{latex}{\out{\hypertarget{method-Engine-run}{}}}
\subsection{\code{Engine$run()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Engine$run(
  entity,
  max_events = 1000,
  max_time = NULL,
  return_observations = TRUE,
  .internal_ctx = NULL
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Engine-run_draw"></a>}}
\if{latex}{\out{\hypertarget{method-Engine-run_draw}{}}}
\subsection{\code{Engine$run_draw()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Engine$run_draw(
  entity,
  params = list(),
  draw_id = 1L,
  sim_id = 1L,
  max_events = 1000,
  max_time = NULL,
  return_observations = TRUE
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Engine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Engine-clone}{}}}
\subsection{\code{Engine$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Engine$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
