% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{PackageProvider}
\alias{PackageProvider}
\title{PackageProvider}
\arguments{
\item{registry}{Named list mapping bundle names to functions returning ModelBundles.}

\item{model_spec}{Named list describing which model/bundle to load.}

\item{n_param_draws}{Integer. Number of global parameter draws to return.}

\item{...}{Additional arguments (reserved for future extension).}
}
\description{
Loads a bundle from in-package functions/objects. Use \code{model_spec$name} to select
among multiple bundle builders. By default, \code{"default"} maps to \code{\link[=default_model_bundle]{default_model_bundle()}}.
}
\section{Fields}{

\describe{
\item{\code{registry}}{Named list mapping bundle names to functions returning ModelBundles.}
}}

\examples{
library(patientSimCore)
prov <- PackageProvider$new()
bundle <- prov$load(list(name = "default"))
names(bundle)

}
