% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{PackageProvider}
\alias{PackageProvider}
\title{PackageProvider}
\arguments{
\item{registry}{Named list mapping bundle names to functions returning ModelBundles.}
}
\description{
Loads a bundle from in-package functions/objects. Use \code{model_spec$name} to select
among multiple bundle builders. By default, \code{"default"} maps to \code{\link[=default_model_bundle]{default_model_bundle()}}.
}
\section{Fields}{

\itemize{
\item \code{registry}: Named list mapping bundle names to functions.
}
}

\section{Methods}{

\itemize{
\item \code{initialize(registry = NULL)}
\item \code{load(model_spec, ...)}
\item \code{sample_param_draws(model_spec, n_param_draws, ...)}
}
}

\examples{
library(patientSimCore)
prov <- PackageProvider$new()
bundle <- prov$load(list(name = "default"))
names(bundle)

}
